backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2
xgboost==1.7.5
requests==2.30.0
joblib==1.2.0
python-dotenv==1.0.0